+++ src/api/index.ts
@@ -1,9 +1,12 @@
 import { IRouter } from 'express'
-import { SignalKMessageHub, WithConfig } from '../app'
+import { SignalKMessageHub, WithConfig, WithFeatures } from '../app'
 import { WithSecurityStrategy } from '../security'
 import { CourseApi } from './course'
+import { FeaturesApi } from './discovery'
 import { ResourcesApi } from './resources'
 import { AutopilotApi } from './autopilot'
+import { SignalKApiId } from '@signalk/server-api'
+
 
 export interface ApiResponse {
   state: 'FAILED' | 'COMPLETED' | 'PENDING'
@@ -38,16 +41,30 @@ export const Responses = {
 }
 
 export const startApis = (
-  app: SignalKMessageHub & WithSecurityStrategy & IRouter & WithConfig
+  app: SignalKMessageHub &
+    WithSecurityStrategy &
+    IRouter &
+    WithConfig &
+    WithFeatures
 ) => {
+  const apiList: Array<SignalKApiId> = []
   const resourcesApi = new ResourcesApi(app)
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   ;(app as any).resourcesApi = resourcesApi
+  apiList.push('resources')
+
   const courseApi = new CourseApi(app, resourcesApi)
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   ;(app as any).courseApi = courseApi
+  apiList.push('course')
+  
   const autopilotApi = new AutopilotApi(app)
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   ;(app as any).autopilotApi = autopilotApi
-  Promise.all([resourcesApi.start(), courseApi.start(), autopilotApi.start()])
+  apiList.push('autopilot')
+
+  const featuresApi = new FeaturesApi(app)
+
+  Promise.all([resourcesApi.start(), courseApi.start(), featuresApi.start(), autopilotApi.start()])
+  return apiList
 }